+++ src/cli/subcommands/mod.rs
@@ -13,7 +13,6 @@ mod car_cmd;
 mod chain_cmd;
 mod config_cmd;
 mod db_cmd;
-mod fetch_params_cmd;
 mod info_cmd;
 mod mpool_cmd;
 mod net_cmd;
@@ -22,7 +21,6 @@ mod shutdown_cmd;
 mod snapshot_cmd;
 mod state_cmd;
 mod sync_cmd;
-mod wallet_cmd;
 
 use std::io::{self, Write};
 
@@ -40,9 +38,9 @@ use tracing::error;
 pub(super) use self::{
     archive_cmd::ArchiveCommands, attach_cmd::AttachCommand, auth_cmd::AuthCommands,
     car_cmd::CarCommands, chain_cmd::ChainCommands, config_cmd::ConfigCommands, db_cmd::DBCommands,
-    fetch_params_cmd::FetchCommands, mpool_cmd::MpoolCommands, net_cmd::NetCommands,
-    send_cmd::SendCommand, shutdown_cmd::ShutdownCommand, snapshot_cmd::SnapshotCommands,
-    state_cmd::StateCommands, sync_cmd::SyncCommands, wallet_cmd::WalletCommands,
+    mpool_cmd::MpoolCommands, net_cmd::NetCommands, send_cmd::SendCommand,
+    shutdown_cmd::ShutdownCommand, snapshot_cmd::SnapshotCommands, state_cmd::StateCommands,
+    sync_cmd::SyncCommands,
 };
 use crate::cli::subcommands::info_cmd::InfoCommand;
 
@@ -57,11 +55,68 @@ pub struct Cli {
     pub cmd: Subcommand,
 }
 
+// This subcommand is hidden and only here to help users migrating to forest-tool
+#[derive(Debug, clap::Args)]
+pub struct FetchCommands {
+    #[arg(short, long)]
+    all: bool,
+    #[arg(short, long)]
+    keys: bool,
+    #[arg(short, long)]
+    dry_run: bool,
+    params_size: Option<String>,
+}
+
+// Those subcommands are hidden and only here to help users migrating to forest-wallet
+#[derive(Debug, clap::Subcommand)]
+pub enum WalletCommands {
+    New {
+        #[arg(default_value = "secp256k1")]
+        signature_type: String,
+    },
+    Balance {
+        address: String,
+    },
+    Default,
+    Export {
+        address: String,
+    },
+    Has {
+        key: String,
+    },
+    Import {
+        path: Option<String>,
+    },
+    List {
+        #[arg(long, alias = "exact-balance", short_alias = 'e')]
+        no_round: bool,
+        #[arg(long, alias = "fixed-unit", short_alias = 'f')]
+        no_abbrev: bool,
+    },
+    SetDefault {
+        key: String,
+    },
+    Sign {
+        #[arg(short)]
+        message: String,
+        #[arg(short)]
+        address: String,
+    },
+    Verify {
+        #[arg(short)]
+        address: String,
+        #[arg(short)]
+        message: String,
+        #[arg(short)]
+        signature: String,
+    },
+}
+
 /// Forest binary sub-commands available.
-#[derive(clap::Subcommand)]
+#[derive(clap::Subcommand, Debug)]
 pub enum Subcommand {
-    /// Download parameters for generating and verifying proofs for given size
-    #[command(name = "fetch-params")]
+    // This subcommand is hidden and only here to help users migrating to forest-tool
+    #[command(hide = true, name = "fetch-params")]
     Fetch(FetchCommands),
 
     /// Interact with Filecoin blockchain
@@ -76,7 +131,8 @@ pub enum Subcommand {
     #[command(subcommand)]
     Net(NetCommands),
 
-    /// Manage wallet
+    // Those subcommands are hidden and only here to help users migrating to forest-wallet
+    #[command(hide = true)]
     #[command(subcommand)]
     Wallet(WalletCommands),
 
@@ -205,7 +261,7 @@ pub(super) fn print_stdout(out: String) {
         .unwrap();
 }
 
-fn prompt_confirm() -> bool {
+pub fn prompt_confirm() -> bool {
     print!("Do you want to continue? [y/n] ");
     std::io::stdout().flush().unwrap();
     let mut line = String::new();