+++ src/cli_shared/cli/mod.rs
@@ -148,7 +148,7 @@ pub struct CliOpts {
 
 impl CliOpts {
     pub fn to_config(&self) -> Result<(Config, Option<ConfigPath>), anyhow::Error> {
-        let path = find_config_path(self);
+        let path = find_config_path(&self.config);
         let mut cfg: Config = match &path {
             Some(path) => {
                 // Read from config file
@@ -231,6 +231,14 @@ impl CliOpts {
     }
 }
 
+/// CLI RPC options
+#[derive(Default, Debug, Parser)]
+pub struct CliRpcOpts {
+    /// Admin token to interact with the node
+    #[arg(long)]
+    pub token: Option<String>,
+}
+
 pub enum ConfigPath {
     Cli(PathBuf),
     Env(PathBuf),
@@ -247,8 +255,8 @@ impl ConfigPath {
     }
 }
 
-fn find_config_path(opts: &CliOpts) -> Option<ConfigPath> {
-    if let Some(s) = &opts.config {
+pub fn find_config_path(config: &Option<String>) -> Option<ConfigPath> {
+    if let Some(s) = config {
         return Some(ConfigPath::Cli(PathBuf::from(s)));
     }
     if let Ok(s) = std::env::var("FOREST_CONFIG_PATH") {